requirements.txt
google-generativeai
flask
flask_session
markdown